Refactor Listbox examples as styleguidist examples
