+++ pages/index.js
@@ -21,7 +21,6 @@ setTimeout(() => {
             <li><a href="./accordions">Accordion</a></li>
             <li><a href="./forms">Forms</a></li>
             <li><a href="./grids">Grid</a></li>
-            <li><a href="./listbox">Listbox</a></li>
             <li><a href="./tabs">Tabs</a></li>
           </ol>
         </nav>
